🚚 Move and rename events component
