+++ apps/cookbook/src/app/showcase/calendar-showcase/calendar-showcase.component.html
@@ -39,5 +39,5 @@
   <cookbook-showcase-properties [properties]="properties"></cookbook-showcase-properties>
 
   <h4>Events:</h4>
-  <cookbook-showcase-events [events]="events"></cookbook-showcase-events>
+  <cookbook-api-description-events [events]="events"></cookbook-api-description-events>
 </div>
