+++ apps/cookbook/src/app/showcase/checkbox-showcase/checkbox-showcase.component.html
@@ -61,5 +61,5 @@
   <cookbook-showcase-properties [properties]="properties"></cookbook-showcase-properties>
 
   <h2>Events</h2>
-  <cookbook-showcase-events [events]="events"></cookbook-showcase-events>
+  <cookbook-api-description-events [events]="events"></cookbook-api-description-events>
 </div>
