+++ apps/cookbook/src/app/showcase/dropdown-showcase/dropdown-showcase.component.html
@@ -88,6 +88,6 @@
 <h2>Properties:</h2>
 <cookbook-showcase-properties [properties]="properties"></cookbook-showcase-properties>
 <h2>Events:</h2>
-<cookbook-showcase-events [events]="events"></cookbook-showcase-events>
+<cookbook-api-description-events [events]="events"></cookbook-api-description-events>
 <h2>Methods:</h2>
 <cookbook-showcase-methods [methods]="methods"></cookbook-showcase-methods>
