+++ apps/cookbook/src/app/showcase/form-field-showcase/form-field-showcase.component.html
@@ -107,7 +107,7 @@
 <h4>Properties:</h4>
 <cookbook-showcase-properties [properties]="sharedProperties"></cookbook-showcase-properties>
 <h4>Events:</h4>
-<cookbook-showcase-events [events]="sharedEvents"></cookbook-showcase-events>
+<cookbook-api-description-events [events]="sharedEvents"></cookbook-api-description-events>
 
 <h2>Input: Specific properties</h2>
 <h4>Properties:</h4>
