+++ apps/cookbook/src/app/showcase/list-load-on-demand-showcase/list-load-on-demand-showcase.component.html
@@ -26,4 +26,4 @@
 <h2>Properties:</h2>
 <cookbook-showcase-properties [properties]="properties"></cookbook-showcase-properties>
 <h2>Events:</h2>
-<cookbook-showcase-events [events]="events"></cookbook-showcase-events>
+<cookbook-api-description-events [events]="events"></cookbook-api-description-events>
