+++ apps/cookbook/src/app/showcase/modal-showcase/modal-showcase.component.html
@@ -246,4 +246,4 @@
 
 <h2>Events:</h2>
 <h3><code>Modal</code></h3>
-<cookbook-showcase-events [events]="events"></cookbook-showcase-events>
+<cookbook-api-description-events [events]="events"></cookbook-api-description-events>
