+++ apps/cookbook/src/app/showcase/page-showcase/page-showcase.component.html
@@ -215,7 +215,7 @@
 <cookbook-showcase-properties [properties]="properties"></cookbook-showcase-properties>
 
 <h2>Page Events:</h2>
-<cookbook-showcase-events [events]="events"></cookbook-showcase-events>
+<cookbook-api-description-events [events]="events"></cookbook-api-description-events>
 
 <h2>Page Layouts:</h2>
 <cookbook-showcase-properties
