+++ apps/cookbook/src/app/showcase/radio-showcase/radio-showcase.component.html
@@ -156,4 +156,4 @@
 
 <h2>Events</h2>
 <h3>Radio Group:</h3>
-<cookbook-showcase-events [events]="groupEvents"></cookbook-showcase-events>
+<cookbook-api-description-events [events]="groupEvents"></cookbook-api-description-events>
