+++ apps/cookbook/src/app/showcase/reorder-list-showcase/reorder-list-showcase.component.html
@@ -12,7 +12,7 @@
   <h2>Properties:</h2>
   <cookbook-showcase-properties [properties]="properties"></cookbook-showcase-properties>
   <h2>Events:</h2>
-  <cookbook-showcase-events [events]="events"></cookbook-showcase-events>
+  <cookbook-api-description-events [events]="events"></cookbook-api-description-events>
   <h2>Methods:</h2>
   <cookbook-showcase-methods [methods]="methods"></cookbook-showcase-methods>
 </div>
