+++ apps/cookbook/src/app/showcase/showcase.module.ts
@@ -6,9 +6,9 @@ import { IonicModule } from '@ionic/angular';
 import { KirbyModule } from '@kirbydesign/designsystem';
 
 import { IphoneModule } from '../iphone/iphone.module';
+import { ShowcaseEventsComponent } from '../shared/api-description/api-description-events/api-description-events.component';
 import { CodeViewerComponent } from '../shared/code-viewer/code-viewer.component';
 import { ExampleViewerComponent } from '../shared/example-viewer/example-viewer.component';
-import { ShowcaseEventsComponent } from '../shared/showcase-events/showcase-events.component';
 import { ShowcaseMethodsComponent } from '../shared/showcase-methods/showcase-methods.component';
 import { ShowcasePropertiesComponent } from '../shared/showcase-properties/showcase-properties.component';
