+++ apps/cookbook/src/app/showcase/slides-showcase/slides-showcase.component.html
@@ -6,5 +6,5 @@
   <cookbook-showcase-properties [properties]="properties"></cookbook-showcase-properties>
   <br />
   <h4>Slides Events:</h4>
-  <cookbook-showcase-events [events]="events"></cookbook-showcase-events>
+  <cookbook-api-description-events [events]="events"></cookbook-api-description-events>
 </div>
