+++ apps/cookbook/src/app/showcase/toolbar-showcase/toolbar-showcase.component.html
@@ -7,5 +7,5 @@
     [properties]="properties"
   ></cookbook-showcase-properties>
   <h2>Events:</h2>
-  <cookbook-showcase-events [events]="events"></cookbook-showcase-events>
+  <cookbook-api-description-events [events]="events"></cookbook-api-description-events>
 </div>
